replace DocSearch with algolia hook
